feat(scribo): update

See merge request olena/pylene!147

+++ pylene/include/scribo/private/span.hpp
@@ -4,8 +4,8 @@ namespace scribo::internal
 {
   struct Span
   {
-    int x;
-    int y;
-    int thickness;
+    float x;
+    float y;
+    int   thickness;
   };
-}
+} // namespace scribo::internal
